Adding Rudimentary Game Menu
This commit adds a game menu. In the game menu the user will set
this cash limit and then start the game. With this change, the
game no longer has a reset button but rather an exit button. When
this is clicked it will bring the user back to menu.